[refactor]: change to float

+++ MAGOI_SEP490/FR.Services/Service/FeedbackService.cs
@@ -65,8 +65,8 @@ namespace FR.Services.Service
 
         public List<Feedback> GetFeedbacks() => _dao.GetFeedbacks();
         public List<Feedback> GetFeedbacksByDate(DateTime startDate, DateTime endDate) => _dao.GetFeedbacksByDate(startDate, endDate);
-        public Double AverageServingScore(List<Feedback> feedbacks) => feedbacks.Count == 0 ? 0 : Math.Round(feedbacks.Select(f => f.ServingScore).Average(), 1);
-        public Double AverageFoodScore(List<Feedback> feedbacks) => feedbacks.Count == 0 ? 0 : Math.Round(feedbacks.Select(f => f.FoodScore).Average(), 1);
+        public float AverageServingScore(List<Feedback> feedbacks) => (float)(feedbacks.Count == 0 ? 0 : Math.Round(feedbacks.Select(f => f.ServingScore).Average(), 1));
+        public float AverageFoodScore(List<Feedback> feedbacks) => (float)(feedbacks.Count == 0 ? 0 : Math.Round(feedbacks.Select(f => f.FoodScore).Average(), 1));
 
     }
 }
